chore (permissions | back): moved `controller/*` > `controllers/*`
